requirements.txt
fastapi
uvicorn[standard]
python-multipart
opencc-python-reimplemented
pysubs2
chardet
